+++ Common/src/main/java/com/almostreliable/morejs/core/Events.java
@@ -3,11 +3,12 @@ package com.almostreliable.morejs.core;
 import com.almostreliable.morejs.features.enchantment.EnchantmentTableServerEventJS;
 import com.almostreliable.morejs.features.enchantment.EnchantmentTableTooltipEventJS;
 import com.almostreliable.morejs.features.enchantment.FilterAvailableEnchantmentsEventJS;
+import com.almostreliable.morejs.features.enchantment.PlayerEnchantEventJS;
 import com.almostreliable.morejs.features.misc.ExperiencePlayerEventJS;
 import com.almostreliable.morejs.features.misc.PiglinPlayerBehaviorEventJS;
 import com.almostreliable.morejs.features.potion.PotionBrewingRegisterEvent;
-import com.almostreliable.morejs.features.structure.StructureLoadEventJS;
 import com.almostreliable.morejs.features.structure.StructureAfterPlaceEventJS;
+import com.almostreliable.morejs.features.structure.StructureLoadEventJS;
 import com.almostreliable.morejs.features.teleport.EntityTeleportsEventJS;
 import com.almostreliable.morejs.features.villager.events.*;
 import dev.latvian.mods.kubejs.event.EventGroup;
@@ -25,7 +26,9 @@ public interface Events {
     EventHandler UPDATE_WANDERER_OFFERS = GROUP.server("updateWandererOffers", () -> UpdateAbstractVillagerOffersEventJS.class);
     EventHandler ENCHANTMENT_TABLE_IS_ENCHANTABLE = GROUP.server("enchantmentTableIsEnchantable", () -> EnchantmentTableServerEventJS.class);
     EventHandler ENCHANTMENT_TABLE_CHANGED = GROUP.server("enchantmentTableChanged", () -> EnchantmentTableServerEventJS.class);
-    EventHandler ENCHANTMENT_TABLE_ENCHANT = GROUP.server("enchantmentTableEnchant", () -> EnchantmentTableServerEventJS.class).hasResult();
+    EventHandler ENCHANTMENT_TABLE_ENCHANT = GROUP
+            .server("enchantmentTableEnchant", () -> PlayerEnchantEventJS.class)
+            .hasResult();
     EventHandler ENCHANTMENT_TABLE_TOOLTIP = GROUP.client("enchantmentTableTooltip", () -> EnchantmentTableTooltipEventJS.class);
     EventHandler TELEPORT = GROUP.server("teleport", () -> EntityTeleportsEventJS.class).hasResult();
     EventHandler STRUCTURE_LOAD = GROUP.server("structureLoad", () -> StructureLoadEventJS.class);